requirements.txt

streamlit>=1.24.0
# LangGraph is optional; fallback included. If you have it, keep it in requirements.
langgraph>=0.2.0
langchain>=0.1.0
langchain-groq>=0.1.0
gitpython>=3.1.31
pytest>=7.4.0
python-dotenv>=1.0.0
sqlalchemy>=2.1.4
